refactor: move argv compatibility helpers
